fix: address PR review comments
- Remove empty filter logic in itemRepository.js
- Fix regex escaping for asterisk character in itemController.js
- Extract formatRandomAttributes to shared utility to avoid code duplication
- Keep test version with i18n for consistency while main function uses raw attributes

Resolves GitHub Copilot PR review suggestions.
