+++ src/app/repositories/itemRepository.js
@@ -112,9 +112,5 @@ exports.getAllById = (ids, filter = {}, sort = {}) => {
     query.orderBy(sort.orderBy, sort.order || "desc");
   }
 
-  if (filter) {
-    // todo
-  }
-
   return query;
 };
